Remove gql templates from spec files
This is necessary because jest collects tests to run from module imports
and when one spec imports another spec, all tests in imported modules
are executed. The fix is to export the gql tags to their own file, and
import those as necessary.
