+++ packages/truffle-db/src/workspace/test/compilation.spec.ts
@@ -1,6 +1,6 @@
-import gql from "graphql-tag";
 import { generateId, Migrations, WorkspaceClient } from './utils';
-import { AddSource } from './source.spec';
+import { AddSource } from './source.graphql';
+import { AddCompilation, GetCompilation, GetAllCompilations } from './compilation.graphql';
 
 describe("Compilation", () => {
   const wsClient = new WorkspaceClient();
@@ -98,88 +98,3 @@ describe("Compilation", () => {
     expect(firstCompilation).toHaveProperty("sources.0.id");
   });
 });
-
-export const GetCompilation = gql`
-  query GetCompilation($id: ID!) {
-    compilation(id: $id) {
-      id
-      compiler {
-        name
-        version
-      }
-      sources {
-        id
-        contents
-      }
-      contracts {
-        source {
-          contents
-        }
-      }
-    }
-  }
-`;
-
-export const GetAllCompilations = gql`
-  query getAllCompilations {
-    compilations {
-      compiler {
-        name
-        version
-        settings
-      }
-
-      sources {
-        id
-      }
-    }
-  }
-`;
-
-export const AddCompilation = gql`
-  mutation AddCompilation($compilerName: String!, $compilerVersion: String!, $sourceId: ID!, $abi:String!) {
-    compilationsAdd(input: {
-      compilations: [{
-        compiler: {
-          name: $compilerName
-          version: $compilerVersion
-        }
-        contracts: [
-        {
-          name:"testing",
-          ast: {
-            json: $abi
-          }
-          source: {
-            id: $sourceId
-          }
-        }]
-        sources: [
-          {
-           id: $sourceId
-          }
-        ]
-      }]
-    }) {
-      compilations {
-        id
-        compiler {
-          name
-        }
-        sources {
-          contents
-        }
-        contracts {
-          source {
-            contents
-            sourcePath
-          }
-          ast {
-            json
-          }
-          name
-        }
-      }
-    }
-  }
-`;
