+++ packages/truffle-db/src/workspace/test/contract.spec.ts
@@ -1,9 +1,9 @@
-import gql from "graphql-tag";
 import { generateId, Migrations, WorkspaceClient } from './utils';
 
-import { AddSource } from './source.spec';
-import { AddBytecode } from './bytecode.spec';
-import { AddCompilation } from './compilation.spec';
+import { AddSource } from './source.graphql';
+import { AddBytecode } from './bytecode.graphql';
+import { AddCompilation } from './compilation.graphql';
+import { AddContracts, GetContract } from './contract.graphql';
 
 describe("Contract", () => {
   let wsClient;
@@ -94,68 +94,3 @@ describe("Contract", () => {
   });
 
 });
-
-export const GetContract = gql`
-  query getContract($id:ID!){
-      contract(id:$id) {
-        name
-        abi {
-          json
-        }
-        sourceContract {
-          source {
-            contents
-          }
-          ast {
-            json
-          }
-        }
-      }
-  }
-`;
-
-
-export const AddContracts = gql`
-  mutation addContracts($contractName: String, $compilationId: ID!, $bytecodeId:ID!, $abi:String!) {
-    contractsAdd(input: {
-      contracts: [{
-        name: $contractName
-        abi: {
-          json: $abi
-        }
-        compilation: {
-          id: $compilationId
-        }
-        sourceContract: {
-          index: 0
-        }
-        constructor: {
-          createBytecode: {
-            id: $bytecodeId
-          }
-        }
-      }]
-    }) {
-      contracts {
-        id
-        name
-        sourceContract {
-          name
-          source {
-            contents
-          }
-          ast {
-            json
-          }
-        }
-        constructor {
-          createBytecode {
-            bytes
-          }
-        }
-      }
-    }
-  }
-`;
-
-
