+++ packages/truffle-db/src/workspace/test/network.graphql.ts
@@ -1,0 +1,38 @@
+import gql from "graphql-tag";
+
+export const GetNetwork = gql`
+  query GetNetwork($id: ID!) {
+    network(id: $id) {
+      networkId
+      id
+    }
+  }
+`;
+
+export const GetAllNetworks = gql`
+  query getAllNetworks {
+    networks  {
+      id
+      networkId
+    }
+  }
+`;
+
+export const AddNetworks = gql`
+  mutation AddNetworks($networkId: NetworkId!, $height: Int!, $hash: String!) {
+    networksAdd(input: {
+      networks: [{
+        networkId: $networkId
+        historicBlock: {
+          height: $height
+          hash: $hash
+        }
+      }]
+    }) {
+      networks {
+        networkId
+        id
+      }
+    }
+  }
+`;
