+++ packages/truffle-db/src/workspace/test/source.graphql.ts
@@ -1,0 +1,27 @@
+import gql from "graphql-tag";
+export const GetSource = gql`
+  query GetSource($id: ID!) {
+    source(id: $id) {
+      id
+      contents
+      sourcePath
+    }
+  }
+`
+
+export const AddSource = gql`
+  mutation AddSource($contents: String!, $sourcePath: String) {
+    sourcesAdd(input: {
+      sources: [
+         {
+           contents: $contents,
+           sourcePath: $sourcePath,
+         }
+      ]
+    }) {
+      sources {
+        id
+      }
+    }
+  }
+`;
