+++ packages/truffle-db/src/workspace/test/source.spec.ts
@@ -1,5 +1,5 @@
-import gql from "graphql-tag";
 import { generateId, Migrations, WorkspaceClient } from './utils';
+import { GetSource, AddSource } from './source.graphql'
 
 describe("Source", () => {
   let wsClient, addSourceResult
@@ -53,31 +53,3 @@ describe("Source", () => {
   });
 
 });
-
-export const GetSource = gql`
-  query GetSource($id: ID!) {
-    source(id: $id) {
-      id
-      contents
-      sourcePath
-    }
-  }
-`
-
-export const AddSource = gql`
-  mutation AddSource($contents: String!, $sourcePath: String) {
-    sourcesAdd(input: {
-      sources: [
-         {
-           contents: $contents,
-           sourcePath: $sourcePath,
-         }
-      ]
-    }) {
-      sources {
-        id
-      }
-    }
-  }
-`
-
